+++ lib/plugin/common/converter/percentage.ts
